Sales-ETL-Dashboard/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.23.0
pandas>=2.1.0
SQLAlchemy>=2.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
